# docs/swagger.yaml
basePath: /
...
host: localhost:9001
info:
  contact: {}
  title: Backend Golang Tickitz App
  version: "1.0"
paths:
...
  /movies:
    get:
      consumes:
      - application/json
      description: Mengambil semua data movie untuk general user. Endpoint ini tidak
        memerlukan otorisasi.
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil semua data movies dengan total count
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found - Tidak ada film yang ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get All Movies
      tags:
      - Movies
  /movies/{movie_id}:
    get:
      consumes:
      - application/json
      description: Mengambil detail film berdasarkan ID
      parameters:
      - description: ID Film
        in: path
        name: movie_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil detail film
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request - movie_id tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found - Film tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Movie Detail
      tags:
      - Movies
  /movies/filter:
    get:
      consumes:
      - application/json
      description: 'Mengambil daftar film dengan filter opsional: judul, genre, dan
        pagination.'
      parameters:
      - description: Filter berdasarkan judul (opsional)
        in: query
        name: title
        type: string
      - description: Filter berdasarkan genre, pisahkan dengan koma jika lebih dari
          satu (opsional)
        in: query
        name: genre
        type: string
      - description: 'Halaman yang ingin diambil (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil daftar film
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found - Tidak ada film yang ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Movies With Filter
      tags:
      - Movies
  /movies/popular:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua film yang sedang populer berdasarkan rating
        atau jumlah penonton
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil daftar popular movies
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found - Tidak ada film popular yang ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Popular Movies
      tags:
      - Movies
  /movies/schedule/{movie_id}:
    get:
      description: Mengambil semua data jadwal film berdasarkan movie ID
      parameters:
      - description: ID Film
        in: path
        name: movie_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID film tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Tidak ada jadwal untuk film tersebut
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get schedules by movie ID
      tags:
      - Movies
  /movies/upcoming:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua film yang akan segera rilis atau upcoming
        movies
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil daftar upcoming movies
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found - Tidak ada film upcoming yang ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Upcoming Movies
      tags:
      - Movies
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"